queue: move docbuilder::queue::BuildQueue to queue::BuildQueue
