+++ src/test/mod.rs
@@ -1,7 +1,7 @@
 mod fakes;
 
 use crate::db::{Pool, PoolConnection};
-use crate::docbuilder::BuildQueue;
+use crate::queue::BuildQueue;
 use crate::storage::s3::TestS3;
 use crate::web::Server;
 use crate::Config;
