+++ src/web/extensions.rs
@@ -1,6 +1,6 @@
 use crate::config::Config;
 use crate::db::Pool;
-use crate::docbuilder::BuildQueue;
+use crate::queue::BuildQueue;
 use crate::web::page::TemplateData;
 use iron::{BeforeMiddleware, IronResult, Request};
 use std::sync::Arc;
